Remove include/poac/io/lockfile.hpp
