+++ include/poac/io.hpp
@@ -1,7 +1,6 @@
 #ifndef POAC_IO_HPP
 #define POAC_IO_HPP
 
-#include <poac/io/lockfile.hpp>
 #include <poac/io/net.hpp>
 #include <poac/io/path.hpp>
 #include <poac/io/tar.hpp>
